dataset/gendata/requirements.txt
datasets>=2.14.0
webdataset>=0.2.0
Pillow>=9.0.0
numpy>=1.21.0

